NewsBot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-ai
botbuilder-dialogs
requests
azure-cosmos